move registration tests into their own folder
