+++ lib/davinci_crd_test_kit/client/v2.0.1/registration/client_registration_verification_test.rb
