+++ lib/davinci_crd_test_kit/client/v2.0.1/registration/client_service_registration_attestation_test.rb
