+++ lib/davinci_crd_test_kit/client/v2.2.1/registration/client_registration_verification_test.rb
