+++ lib/davinci_crd_test_kit/client/v2.2.1/registration/client_service_registration_attestation_test.rb
